+++ CHANGELOG
@@ -1160,7 +1160,7 @@ Resolved issues:
 	#506	added custom info.plist containing the path to OMSSA/XTandem for the GUI tools
 	#505	goto dialogs now act with error tolerance
 	#504	Fixed XTandemXMLFile loading when no encoding is given by XTandemXML file by enforcing ISO-Latin-1 (default would be utf-8, breaking the aminoacid sequence readin). Introduced enforceEncoding(encoding) to XMLFile.
-	#316	SpectraViewWidget is not cleaned up when last layer is deleted
+	#316	SpectraTreeTab is not cleaned up when last layer is deleted
 	#283	TOPPView: Scan view widget does not clear after closing last file
 	#379	MapAlignerPoseClustering: -reference:index has no effect
 	#384	replace zip libraries with The Boost Iostreams Library (already in contrib)
