+++ src/openms_gui/include/OpenMS/VISUAL/sources.cmake
@@ -35,7 +35,7 @@ Plot3DWidget.h
 PlotCanvas.h
 PlotWidget.h
 RecentFilesMenu.h
-SpectraViewWidget.h
+SpectraTreeTab.h
 SpectraIDViewTab.h
 SwathLibraryStats.h
 TableView.h
